feat(theme): update Poimandres theme to v1.0.14
This commit includes several changes to the Poimandres theme. The version
has been updated from 1.0.13 to 1.0.14. The color palette has been
muted and the border radii have been increased from 0px to 4px. The
scrollbar width has been increased from 0px to 8px. The padding for the
menu has been adjusted and the border radius for several elements has
been set to 0. This commit aims to improve the overall aesthetics of the
theme.